requirements.txt
streamlit>=1.28.0
haystack-ai>=2.0.0
transformers>=4.21.0
torch>=2.0.0
accelerate>=0.20.0
serpapi>=0.1.5
beautifulsoup4>=4.12.0
requests>=2.28.0
pandas>=1.5.0
numpy>=1.21.0
sentencepiece>=0.1.99
tqdm>=4.64.0